move overlay to custom PopupWrapper
