+++ apps/fishing-map/features/map/overlays/draw/CoordinateEditOverlay.tsx
@@ -1,16 +1,14 @@
-import { HtmlOverlay, HtmlOverlayItem } from '@nebula.gl/overlays'
 import { useTranslation } from 'react-i18next'
 import { useCallback, useState } from 'react'
 import { useSelector } from 'react-redux'
 import { Button, IconButton, InputText } from '@globalfishingwatch/ui-components'
-import { useMapViewport } from 'features/map/map-viewport.hooks'
 import { selectMapDrawingMode } from 'routes/routes.selectors'
+import PopupWrapper from 'features/map/popups/PopupWrapper'
 import { useDrawLayerInstance } from './draw.hooks'
 import styles from './DrawDialog.module.css'
 
 export const CoordinateEditOverlay = () => {
   const { t } = useTranslation()
-  const viewport = useMapViewport()
   const drawLayer = useDrawLayerInstance()
   const drawingMode = useSelector(selectMapDrawingMode)
 
@@ -87,55 +85,52 @@ export const CoordinateEditOverlay = () => {
     // }
   }, [])
 
-  return currentPointCoordinates?.length ? (
-    <div onPointerUp={(event) => event.preventDefault()} className={styles.popup}>
-      <HtmlOverlay viewport={viewport} key="2">
-        <HtmlOverlayItem
-          style={{ pointerEvents: 'all', transform: 'translate(-50%,-105%)' }}
-          coordinates={[Number(editingPointLongitude), Number(editingPointLatitude)]}
-        >
-          <div className={styles.popupContent}>
-            <div className={styles.flex}>
-              <InputText
-                step="0.01"
-                type="number"
-                value={editingPointLatitude}
-                label={t('common.latitude', 'Latitude')}
-                onChange={onHandleLatitudeChange}
-                className={styles.shortInput}
-              />
-              <InputText
-                step="0.01"
-                type="number"
-                className={styles.shortInput}
-                value={editingPointLongitude}
-                label={t('common.longitude', 'longitude')}
-                onChange={onHandleLongitudeChange}
-              />
-            </div>
-            <div className={styles.popupButtons}>
-              <IconButton
-                icon="delete"
-                type="warning-border"
-                onClick={onDeletePoint}
-                disabled={false}
-                tooltip={
-                  allowDeletePoint
-                    ? t('layer.removePoint', 'Remove point')
-                    : t('layer.removePointNotAllowed', 'Geometry needs at least 3 points')
-                }
-              />
-              <Button
-                disabled={editingPointLatitude === null || editingPointLongitude === null}
-                onClick={onConfirm}
-                className={styles.confirmBtn}
-              >
-                {t('common.confirm', 'Confirm')}
-              </Button>
-            </div>
-          </div>
-        </HtmlOverlayItem>
-      </HtmlOverlay>
-    </div>
-  ) : null
+  if (!currentPointCoordinates?.length) {
+    return null
+  }
+
+  return (
+    <PopupWrapper latitude={editingPointLongitude} longitude={editingPointLatitude}>
+      <div className={styles.popupContent}>
+        <div className={styles.flex}>
+          <InputText
+            step="0.01"
+            type="number"
+            value={editingPointLatitude}
+            label={t('common.latitude', 'Latitude')}
+            onChange={onHandleLatitudeChange}
+            className={styles.shortInput}
+          />
+          <InputText
+            step="0.01"
+            type="number"
+            className={styles.shortInput}
+            value={editingPointLongitude}
+            label={t('common.longitude', 'longitude')}
+            onChange={onHandleLongitudeChange}
+          />
+        </div>
+        <div className={styles.popupButtons}>
+          <IconButton
+            icon="delete"
+            type="warning-border"
+            onClick={onDeletePoint}
+            disabled={false}
+            tooltip={
+              allowDeletePoint
+                ? t('layer.removePoint', 'Remove point')
+                : t('layer.removePointNotAllowed', 'Geometry needs at least 3 points')
+            }
+          />
+          <Button
+            disabled={editingPointLatitude === null || editingPointLongitude === null}
+            onClick={onConfirm}
+            className={styles.confirmBtn}
+          >
+            {t('common.confirm', 'Confirm')}
+          </Button>
+        </div>
+      </div>
+    </PopupWrapper>
+  )
 }
