+++ apps/fishing-map/features/map/overlays/draw/DrawDialog.module.css
@@ -67,13 +67,10 @@
   margin-right: var(--space-M);
 }
 
-.popup {
+.popupContent {
   color: var(--color-primary-blue);
   box-shadow: var(--box-shadow);
   max-width: 40vw;
-}
-
-.popup .popupContent {
   background: var(--color-white);
   padding: var(--space-S);
 }
