+++ apps/fishing-map/features/map/popups/MapPopups.tsx
@@ -4,6 +4,8 @@ import { useMapHoverInteraction } from '@globalfishingwatch/deck-layer-composer'
 import PopupWrapper from 'features/map/popups/PopupWrapper'
 import { selectClickedEvent } from '../map.slice'
 import { useClickedEventConnect } from '../map-interactions.hooks'
+import PopupByCategory from './PopupByCategory'
+import styles from './Popup.module.css'
 
 function MapPopups() {
   const hoverInteraction = useMapHoverInteraction()
@@ -19,10 +21,25 @@ function MapPopups() {
   return (
     <Fragment>
       {hoverInteraction && !clickInteraction?.features?.length && (
-        <PopupWrapper interaction={hoverInteraction} type="hover" />
+        <PopupWrapper
+          latitude={hoverInteraction.latitude}
+          longitude={hoverInteraction.longitude}
+          className={styles.hover}
+          showArrow={false}
+          showClose={false}
+        >
+          <PopupByCategory interaction={hoverInteraction} type="hover" />
+        </PopupWrapper>
       )}
       {clickInteraction && (
-        <PopupWrapper interaction={clickInteraction} type="click" onClose={closePopup} />
+        <PopupWrapper
+          latitude={clickInteraction.latitude}
+          longitude={clickInteraction.longitude}
+          className={styles.click}
+          onClose={closePopup}
+        >
+          <PopupByCategory interaction={clickInteraction} type="click" />
+        </PopupWrapper>
       )}
     </Fragment>
   )
