+++ apps/fishing-map/features/map/popups/Popup.module.css
@@ -1,6 +1,5 @@
 .popup {
   font: var(--font-S);
-  pointer-events: none;
   z-index: 2;
 }
 
@@ -13,7 +12,6 @@
   width: 14px;
   height: 14px;
   z-index: 1;
-  pointer-events: none;
   background: var(--color-white);
   transform: rotate(45deg);
 }
@@ -39,8 +37,8 @@
   justify-content: center;
 }
 
-.popup.click {
-  pointer-events: all;
+.popup.hover {
+  pointer-events: none;
 }
 
 .hover .content {
