+++ apps/fishing-map/features/map/popups/PopupWrapper.tsx
@@ -1,7 +1,5 @@
-import { Fragment, useRef } from 'react'
+import { useRef } from 'react'
 import cx from 'classnames'
-import { groupBy, uniqBy } from 'lodash'
-import { useSelector } from 'react-redux'
 import {
   offset,
   autoPlacement,
@@ -12,43 +10,11 @@ import {
   arrow,
   FloatingArrow,
 } from '@floating-ui/react'
-import { DataviewCategory, DataviewType } from '@globalfishingwatch/api-types'
-import { IconButton, Spinner } from '@globalfishingwatch/ui-components'
+import { IconButton } from '@globalfishingwatch/ui-components'
 import { InteractionEvent } from '@globalfishingwatch/deck-layer-composer'
-import {
-  ContextPickingObject,
-  FourwingsComparisonMode,
-  FourwingsHeatmapPickingObject,
-  FourwingsPositionsPickingObject,
-  RulerPickingObject,
-  UserLayerPickingObject,
-  VesselEventPickingObject,
-} from '@globalfishingwatch/deck-layers'
-import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
-import { POPUP_CATEGORY_ORDER } from 'data/config'
-import DetectionsTooltipRow from 'features/map/popups/DetectionsLayers'
-import UserPointsTooltipSection from 'features/map/popups/UserPointsLayers'
-import { AsyncReducerStatus } from 'utils/async-slice'
 import { useMapViewport } from 'features/map/map-viewport.hooks'
-import { getDatasetTitleByDataview } from 'features/datasets/datasets.utils'
-import { selectAllDataviewInstancesResolved } from 'features/dataviews/selectors/dataviews.instances.selectors'
-import ComparisonRow from 'features/map/popups/ComparisonRow'
-import WorkspacePointsTooltipSection from 'features/map/popups/WorkspacePointsLayers'
-import {
-  SliceExtendedFourwingsPickingObject,
-  selectApiEventStatus,
-  selectFishingInteractionStatus,
-} from '../map.slice'
 import { MAP_WRAPPER_ID } from '../map.config'
-import UserContextTooltipSection from './UserContextLayers'
 import styles from './Popup.module.css'
-import ActivityTooltipRow from './ActivityLayers'
-import TileClusterTooltipRow from './TileClusterTooltipRow'
-import ContextTooltipSection from './ContextLayers'
-import VesselEventsLayers from './VesselEventsLayers'
-import EnvironmentTooltipSection from './EnvironmentLayers'
-import PositionsRow from './PositionsRow'
-import RulerTooltip from './RulerTooltip'
 
 const overflowMiddlware: Middleware = {
   name: 'overflow',
@@ -71,23 +37,26 @@ const overflowMiddlware: Middleware = {
 }
 
 type PopupWrapperProps = {
-  interaction: InteractionEvent | null
-  closeButton?: boolean
-  closeOnClick?: boolean
+  latitude: InteractionEvent['latitude'] | null
+  longitude: InteractionEvent['longitude'] | null
+  showArrow?: boolean
+  showClose?: boolean
   className?: string
   onClose?: () => void
-  type?: 'hover' | 'click'
+  children: React.ReactNode
 }
 
-const OMITED_CATEGORIES = ['draw']
-
-function PopupWrapper({ interaction, type = 'hover', className = '', onClose }: PopupWrapperProps) {
+function PopupWrapper({
+  latitude,
+  longitude,
+  showArrow = true,
+  showClose = true,
+  className = '',
+  onClose,
+  children,
+}: PopupWrapperProps) {
   // Assuming only timeComparison heatmap is visible, so timerange description apply to all
   const mapViewport = useMapViewport()
-  const dataviews = useSelector(selectAllDataviewInstancesResolved) as UrlDataviewInstance[]
-
-  const activityInteractionStatus = useSelector(selectFishingInteractionStatus)
-  const apiEventStatus = useSelector(selectApiEventStatus)
 
   const arrowRef = useRef<SVGSVGElement>(null)
   const { refs, floatingStyles, context } = useFloating({
@@ -103,267 +72,23 @@ function PopupWrapper({ interaction, type = 'hover', className = '', onClose }: 
     ],
   })
 
-  if (!mapViewport || !interaction || !interaction.features?.length) return null
-
-  const visibleFeatures = interaction?.features.filter(
-    (feature) => !OMITED_CATEGORIES.includes(feature.category)
-  )
-
-  if (!visibleFeatures.length) return null
-
-  const [left, top] = mapViewport.project([interaction.longitude, interaction.latitude])
-
-  const featureByCategory = groupBy(
-    visibleFeatures
-      // Needed to create a new array and not muting with sort
-      .map((feature) => feature)
-      .sort(
-        (a, b) =>
-          POPUP_CATEGORY_ORDER.indexOf(a?.category as DataviewCategory) -
-          POPUP_CATEGORY_ORDER.indexOf(b?.category as DataviewCategory)
-      ),
-    'category'
-  )
+  if (!mapViewport || !latitude || !longitude) return null
+  const [left, top] = mapViewport.project([longitude, latitude])
 
   return (
     <div
       ref={refs.setReference}
-      style={{ position: 'absolute', top, left }}
-      className={cx(styles.popup, styles[type], className)}
+      style={{ position: 'absolute', top, left, zIndex: 2 }}
+      className={cx(styles.popup, className)}
     >
       <div className={styles.contentWrapper} ref={refs.setFloating} style={floatingStyles}>
-        {type === 'click' && <FloatingArrow fill="white" ref={arrowRef} context={context} />}
-        {type === 'click' && onClose !== undefined && (
+        {showArrow && <FloatingArrow fill="white" ref={arrowRef} context={context} />}
+        {showClose && onClose !== undefined && (
           <div className={styles.close}>
             <IconButton type="invert" size="small" icon="close" onClick={onClose} />
           </div>
         )}
-        <div className={styles.content}>
-          {Object.entries(featureByCategory)?.map(([featureCategory, features]) => {
-            switch (featureCategory) {
-              // TODO: deck restore this popup
-              // case DataviewCategory.Comparison:
-              //   return (
-              //     <ComparisonRow
-              //       key={featureCategory}
-              //       feature={features[0]}
-              //       showFeaturesDetails={type === 'click'}
-              //     />
-              //   )
-              case DataviewCategory.Activity: {
-                const positionFeatures = (features as SliceExtendedFourwingsPickingObject[]).filter(
-                  (feature) => feature.visualizationMode === 'positions'
-                )
-                const uniqPositionFeatures = uniqBy(positionFeatures, 'properties.id')
-                const heatmapFeatures = (features as SliceExtendedFourwingsPickingObject[]).filter(
-                  (feature) => feature.visualizationMode !== 'positions'
-                )
-
-                return [...uniqPositionFeatures, ...heatmapFeatures].map((feature, i) => {
-                  if (feature.visualizationMode === 'positions') {
-                    return (
-                      <PositionsRow
-                        key={feature.id}
-                        feature={feature as any as FourwingsPositionsPickingObject}
-                        showFeaturesDetails={type === 'click'}
-                      />
-                    )
-                  }
-                  return feature.sublayers.map((sublayer, j) => {
-                    const dataview = dataviews.find((d) => d.id === sublayer.id)
-                    return feature.comparisonMode === FourwingsComparisonMode.TimeCompare ? (
-                      <ComparisonRow
-                        key={featureCategory}
-                        feature={features[0] as FourwingsHeatmapPickingObject}
-                        showFeaturesDetails={type === 'click'}
-                      />
-                    ) : (
-                      <ActivityTooltipRow
-                        key={`${i}-${j}`}
-                        loading={activityInteractionStatus === AsyncReducerStatus.Loading}
-                        feature={{
-                          ...sublayer,
-                          category: feature.category as DataviewCategory,
-                          title: dataview
-                            ? getDatasetTitleByDataview(dataview, { showPrivateIcon: false })
-                            : feature.title,
-                        }}
-                        showFeaturesDetails={type === 'click'}
-                      />
-                    )
-                  })
-                })
-              }
-              case DataviewCategory.Detections: {
-                return (features as SliceExtendedFourwingsPickingObject[])?.map((feature, i) => {
-                  return feature.sublayers.map((sublayer, j) => {
-                    const dataview = dataviews.find((d) => d.id === sublayer.id)
-                    return (
-                      <DetectionsTooltipRow
-                        key={`${i}-${j}`}
-                        loading={activityInteractionStatus === AsyncReducerStatus.Loading}
-                        feature={{
-                          ...sublayer,
-                          category: feature.category as DataviewCategory,
-                          title: dataview
-                            ? getDatasetTitleByDataview(dataview, { showPrivateIcon: false })
-                            : feature.title,
-                        }}
-                        showFeaturesDetails={type === 'click'}
-                      />
-                    )
-                  })
-                })
-              }
-              case DataviewCategory.Events: {
-                if (apiEventStatus === AsyncReducerStatus.Loading) {
-                  return (
-                    <div className={styles.loading}>
-                      <Spinner size="small" />
-                    </div>
-                  )
-                }
-                return (
-                  <TileClusterTooltipRow
-                    key={featureCategory}
-                    // TODO:deck review typings here
-                    features={features as any}
-                    showFeaturesDetails={type === 'click'}
-                  />
-                )
-              }
-              // TODO: deck restore this popup
-              case DataviewCategory.Environment: {
-                // TODO:deck review if this is needed
-                //   const contextEnvironmentalFeatures = features.filter(
-                //     (feature) =>
-                //       feature.type === DataviewType.Context ||
-                //       feature.type === DataviewType.UserContext
-                //   )
-                // const environmentalFeatures = features.filter(
-                //   (feature) =>
-                //     feature.type !== DataviewType.Context &&
-                //     feature.type !== DataviewType.UserContext
-                // )
-                return (
-                  <Fragment key={featureCategory}>
-                    {/* <ContextTooltipSection
-                      features={contextEnvironmentalFeatures}
-                      showFeaturesDetails={type === 'click'}
-                    /> */}
-                    <EnvironmentTooltipSection
-                      features={features as SliceExtendedFourwingsPickingObject[]}
-                      showFeaturesDetails={type === 'click'}
-                    />
-                  </Fragment>
-                )
-              }
-              case DataviewCategory.Context: {
-                const contextFeatures = (features as ContextPickingObject[]).filter(
-                  (feature) => feature.subcategory !== DataviewType.UserPoints
-                )
-                const pointFeatures = (features as UserLayerPickingObject[]).filter(
-                  (feature) => feature.subcategory === DataviewType.UserPoints
-                )
-                // const defaultContextFeatures = features.filter(
-                //   (feature) => feature.type === DataviewType.Context
-                // )
-                // const workspacePointsFeatures = features.filter(
-                //   (feature) => feature.source === WORKSPACE_GENERATOR_ID
-                // )
-                // const areaBufferFeatures = features.filter(
-                //   (feature) => feature.source === REPORT_BUFFER_GENERATOR_ID
-                // )
-                // const annotationFeatures = (features as MapAnnotation[]).filter(
-                //   (feature) => feature.category === 'annotations'
-                // )
-                // Workaround to show user context features in the context section
-                const userContextFeatures = (features as UserLayerPickingObject[]).filter(
-                  (feature) => feature.subcategory === DataviewType.UserContext
-                )
-                return (
-                  <Fragment key={featureCategory}>
-                    {/* <AnnotationTooltip features={annotationFeatures} /> */}
-                    {/* <ReportBufferTooltip features={areaBufferFeatures} /> */}
-                    <UserPointsTooltipSection
-                      features={pointFeatures}
-                      showFeaturesDetails={type === 'click'}
-                    />
-                    <ContextTooltipSection
-                      features={contextFeatures}
-                      showFeaturesDetails={type === 'click'}
-                    />
-                    <UserContextTooltipSection
-                      features={userContextFeatures}
-                      showFeaturesDetails={type === 'click'}
-                    />
-                  </Fragment>
-                )
-              }
-              // TODO: deck restore this popup
-              case DataviewCategory.User: {
-                const userPointFeatures = (features as UserLayerPickingObject[]).filter(
-                  (feature) =>
-                    feature.subcategory === DataviewType.UserPoints ||
-                    feature.subcategory === 'draw-points'
-                )
-                const userContextFeatures = (features as UserLayerPickingObject[]).filter(
-                  (feature) =>
-                    feature.subcategory === DataviewType.UserContext ||
-                    feature.subcategory === 'draw-polygons'
-                )
-                return (
-                  <Fragment key={featureCategory}>
-                    <UserPointsTooltipSection
-                      features={userPointFeatures}
-                      showFeaturesDetails={type === 'click'}
-                    />
-                    <UserContextTooltipSection
-                      features={userContextFeatures}
-                      showFeaturesDetails={type === 'click'}
-                    />
-                  </Fragment>
-                )
-              }
-
-              case DataviewCategory.Vessels: {
-                return (
-                  <VesselEventsLayers
-                    key={featureCategory}
-                    features={features as VesselEventPickingObject[]}
-                    showFeaturesDetails={type === 'click'}
-                  />
-                )
-              }
-
-              case DataviewCategory.Workspaces: {
-                return (
-                  <WorkspacePointsTooltipSection
-                    key={featureCategory}
-                    features={features as any}
-                    showFeaturesDetails={type === 'click'}
-                  />
-                )
-              }
-
-              case 'rulers': {
-                const rulersFeatures = (features as RulerPickingObject[]).filter(
-                  (f) => f.properties.order === 'start' || f.properties.order === 'end'
-                )
-                return (
-                  <RulerTooltip
-                    key={featureCategory}
-                    features={rulersFeatures}
-                    showFeaturesDetails={type === 'click'}
-                  />
-                )
-              }
-
-              default:
-                return null
-            }
-          })}
-        </div>
+        <div className={styles.content}>{children}</div>
       </div>
     </div>
   )
